Change specs to support both single and multiple accounts
