+++ paypal_express.yml
@@ -1,5 +1,4 @@
 :paypal_express:
-  - :account_id: paypal_test_account
     :signature: <%= ENV['SIGNATURE'] %>
     :login: <%= ENV['LOGIN'] %>
     :password: <%= ENV['PASSWORD'] %>
