+++ spec/paypal_express/remote/baid_spec.rb
@@ -1,53 +1,12 @@
 require 'spec_helper'
+require_relative 'build_plugin_helpers'
+require_relative 'baid_helpers'
 
 ActiveMerchant::Billing::Base.mode = :test
 
-describe Killbill::PaypalExpress::PaymentPlugin do
-
-  include ::Killbill::Plugin::ActiveMerchant::RSpec
-
-  # Share the BAID
-  before(:all) do
-    @plugin = build_plugin(::Killbill::PaypalExpress::PaymentPlugin, 'paypal_express')
-    svcs = @plugin.kb_apis.proxied_services
-    svcs[:payment_api] = PaypalExpressJavaPaymentApi.new(@plugin)
-    @plugin.kb_apis = ::Killbill::Plugin::KillbillApi.new('paypal_express', svcs)
-    @plugin.start_plugin
-
-    @call_context = build_call_context
-
-    @properties = []
-    @amount = BigDecimal.new('100')
-    @currency = 'USD'
-
-    kb_account_id = SecureRandom.uuid
-    external_key, kb_account_id = create_kb_account(kb_account_id, @plugin.kb_apis.proxied_services[:account_user_api])
-
-    # Initiate the setup process
-    response = create_token(kb_account_id, @call_context.tenant_id)
-    token = response.token
-    print "\nPlease go to #{@plugin.to_express_checkout_url(response, @call_context.tenant_id)} to proceed and press any key to continue...
-Note: you need to log-in with a paypal sandbox account (create one here: https://developer.paypal.com/webapps/developer/applications/accounts)\n"
-    $stdin.gets
-
-    # Complete the setup process
-    @properties << build_property('token', token)
-    @pm = create_payment_method(::Killbill::PaypalExpress::PaypalExpressPaymentMethod, kb_account_id, @call_context.tenant_id, @properties)
-
-    # Verify our table directly. Note that @pm.token is the baid
-    payment_methods = ::Killbill::PaypalExpress::PaypalExpressPaymentMethod.from_kb_account_id_and_token(@pm.token, kb_account_id, @call_context.tenant_id)
-    payment_methods.size.should == 1
-    payment_method = payment_methods.first
-    payment_method.should_not be_nil
-    payment_method.paypal_express_payer_id.should_not be_nil
-    payment_method.token.should == @pm.token
-    payment_method.kb_payment_method_id.should == @pm.kb_payment_method_id
-  end
-
+shared_examples 'baid_spec_common' do
   before(:each) do
     ::Killbill::PaypalExpress::PaypalExpressTransaction.delete_all
-    ::Killbill::PaypalExpress::PaypalExpressResponse.delete_all
-
     kb_payment_id = SecureRandom.uuid
     1.upto(6) do
       @kb_payment = @plugin.kb_apis.proxied_services[:payment_api].add_payment(kb_payment_id)
@@ -70,6 +29,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     # Try a full refund
     refund_response = @plugin.refund_payment(@pm.kb_account_id, @kb_payment.id, @kb_payment.transactions[1].id, @pm.kb_payment_method_id, @amount, @currency, @properties, @call_context)
@@ -87,6 +47,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
     payment_infos[1].kb_payment_id.should == @kb_payment.id
     payment_infos[1].transaction_type.should == :REFUND
     payment_infos[1].amount.should == @amount
@@ -94,6 +55,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[1].status.should == :PROCESSED
     payment_infos[1].gateway_error.should == 'Success'
     payment_infos[1].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[1].properties, 'payment_processor_account_id').should == @payment_processor_account_id
   end
 
   it 'should be able to auth, capture and refund' do
@@ -112,6 +74,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     # Try multiple partial captures
     partial_capture_amount = BigDecimal.new('10')
@@ -131,6 +94,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
       payment_infos[i].status.should == :PROCESSED
       payment_infos[i].gateway_error.should == 'Success'
       payment_infos[i].gateway_error_code.should be_nil
+      find_value_from_properties(payment_infos[i].properties, 'payment_processor_account_id').should == @payment_processor_account_id
     end
 
     # Try a partial refund
@@ -149,6 +113,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[4].status.should == :PROCESSED
     payment_infos[4].gateway_error.should == 'Success'
     payment_infos[4].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[4].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     # Try to capture again
     payment_response = @plugin.capture_payment(@pm.kb_account_id, @kb_payment.id, @kb_payment.transactions[5].id, @pm.kb_payment_method_id, partial_capture_amount, @currency, @properties, @call_context)
@@ -166,6 +131,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[5].status.should == :PROCESSED
     payment_infos[5].gateway_error.should == 'Success'
     payment_infos[5].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[5].properties, 'payment_processor_account_id').should == @payment_processor_account_id
   end
 
   it 'should be able to auth and void' do
@@ -184,6 +150,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     payment_response = @plugin.void_payment(@pm.kb_account_id, @kb_payment.id, @kb_payment.transactions[1].id, @pm.kb_payment_method_id, @properties, @call_context)
     payment_response.status.should eq(:PROCESSED), payment_response.gateway_error
@@ -199,6 +166,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
     payment_infos[1].kb_payment_id.should == @kb_payment.id
     payment_infos[1].transaction_type.should == :VOID
     payment_infos[1].amount.should be_nil
@@ -206,6 +174,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[1].status.should == :PROCESSED
     payment_infos[1].gateway_error.should == 'Success'
     payment_infos[1].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[1].properties, 'payment_processor_account_id').should == @payment_processor_account_id
   end
 
   it 'should be able to auth, partial capture and void' do
@@ -224,6 +193,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[0].status.should == :PROCESSED
     payment_infos[0].gateway_error.should == 'Success'
     payment_infos[0].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[0].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     partial_capture_amount = BigDecimal.new('10')
     payment_response = @plugin.capture_payment(@pm.kb_account_id, @kb_payment.id, @kb_payment.transactions[1].id, @pm.kb_payment_method_id, partial_capture_amount, @currency, @properties, @call_context)
@@ -241,6 +211,7 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[1].status.should == :PROCESSED
     payment_infos[1].gateway_error.should == 'Success'
     payment_infos[1].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[1].properties, 'payment_processor_account_id').should == @payment_processor_account_id
 
     payment_response = @plugin.void_payment(@pm.kb_account_id, @kb_payment.id, @kb_payment.transactions[2].id, @pm.kb_payment_method_id, @properties, @call_context)
     payment_response.status.should eq(:PROCESSED), payment_response.gateway_error
@@ -256,32 +227,58 @@ Note: you need to log-in with a paypal sandbox account (create one here: https:/
     payment_infos[2].status.should == :PROCESSED
     payment_infos[2].gateway_error.should == 'Success'
     payment_infos[2].gateway_error_code.should be_nil
+    find_value_from_properties(payment_infos[2].properties, 'payment_processor_account_id').should == @payment_processor_account_id
   end
 
   it 'should generate forms correctly' do
     context = @plugin.kb_apis.create_context(@call_context.tenant_id)
-    fields  = @plugin.hash_to_properties(
-                 :order_id => '1234',
-                 :amount   => 12,
-              )
+    fields = @plugin.hash_to_properties(
+        :order_id => '1234',
+        :amount => 12,
+    )
 
     properties = @plugin.hash_to_properties(
-      :create_pending_payment => false
+        :create_pending_payment => false
     )
 
     form = @plugin.build_form_descriptor(@pm.kb_account_id, fields, properties, context)
 
     form.kb_account_id.should == @pm.kb_account_id
-    form.form_method.should   == 'GET'
+    form.form_method.should == 'GET'
     form.form_url.should start_with('https://www.sandbox.paypal.com/cgi-bin/webscr')
   end
 
-  private
+end
 
-  def create_token(kb_account_id, kb_tenant_id)
-    private_plugin = ::Killbill::PaypalExpress::PrivatePaymentPlugin.new
-    response       = private_plugin.initiate_express_checkout(kb_account_id, kb_tenant_id)
-    response.success.should be_true
-    response
+describe Killbill::PaypalExpress::PaymentPlugin do
+
+  include ::Killbill::Plugin::ActiveMerchant::RSpec
+  include BuildPluginHelpers
+  include BaidHelpers
+
+  context 'baid test with a single account' do
+    # Share the BAID
+    before(:all) do
+      # delete once here because we need to keep the initial response for later tests to find the payment processor account id
+      ::Killbill::PaypalExpress::PaypalExpressResponse.delete_all
+      @payment_processor_account_id = 'default'
+      @plugin = build_start_paypal_plugin
+      baid_setup
+    end
+
+    include_examples 'baid_spec_common'
+  end
+
+  context 'baid tests with multiple accounts' do
+    # Share the BAID
+    before(:all) do
+      # delete once here because we need to keep the initial response for later tests to find the payment processor account id
+      ::Killbill::PaypalExpress::PaypalExpressResponse.delete_all
+      @payment_processor_account_id = 'paypal_test_account'
+      @plugin = build_start_paypal_plugin @payment_processor_account_id
+      baid_setup @payment_processor_account_id
+    end
+
+    include_examples 'baid_spec_common'
   end
 end
